pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trading-bot"
version = "1.0.0"
description = "Binance Futures Testnet Trading Bot — Python Developer Assessment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "python-dotenv>=1.0.1",
]

[project.scripts]
trading-bot = "bot.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
